backend_api/pyproject.toml
[tool.poetry]
name = "autodeploy"
version = "0.1.0"
description = ""
authors = ["Guilherme Rodrigues <guilherme.b.rodrigues@skf.com>"]

[tool.poetry.dependencies]
python = "^3.9.10"
fastapi = "^0.75.2"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
SQLAlchemy = "^1.4.35"
httpx = "^0.22.0"
grpcio = "^1.45.0"
grpcio-tools = "^1.45.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
isort = "^5.10.1"
prospector = "^1.7.7"
blue = "^0.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
